# lib/api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a REST server for ioBroker."
  version: "2.0.3"
  title: "ioBroker Swagger UI"
  contact:
    email: "admin@iobroker.net"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/v1"
...
securityDefinitions:
  basicAuth:
    type: basic
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter your bearer token in the format **Bearer &lt;token&gt;**"
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /sendto/{instance}:
    x-swagger-router-controller: sendTo
    get:
      tags:
        - "sendTo"
      summary: "Send message to instance"
      description: "You can send message to instance that supports it"
      operationId: "sendTo"
      produces:
        - "application/json"
      parameters:
        - name: "instance"
          in: "path"
          description: "Instance, like javascript.0"
          required: true
          type: "string"
        - name: "message"
          in: "query"
          description: "Command name, like toScript"
          type: "string"
          required: true
        - in: "query"
          name: "data"
          description: "Data to be sent to instance. JSON or string"
          required: false
          type: "string"
        - in: "query"
          name: "noResponse"
          description: "Do not wait for answer or no answer"
          required: false
          type: "boolean"
        - in: "query"
          name: "timeout"
          description: "Timeout for answer"
          required: false
          type: "number"
      responses:
        200:
          description: "Answer from instance if received"
        408:
          description: "Timeout"
        422:
          description: "Invalid state ID supplied"
        500:
          description: "instance is offline"
    post:
      tags:
        - "sendTo"
      summary: "Send message to instance as POST"
      description: "You can send message to instance that supports it"
      operationId: "sendToPost"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "instance"
          in: "path"
          description: "Instance, like javascript.0"
          required: true
          type: "string"
        - name: "data"
          in: "body"
          description: "Data and message"
          required: true
          schema:
            $ref: "#/definitions/SendToData"
      responses:
        200:
          description: "Answer from instance if received"
        408:
          description: "Timeout"
        422:
          description: "Invalid state ID supplied"
        500:
          description: "instance is offline"
...
definitions:
...
  SendToData:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Message or command, like toScript"
      data:
        type: "object"
        description: "String or JSON Object that will be sent"
      noAnswer:
        type: "boolean"
        description: "If wait for answer"
      timeout:
        type: "number"
        description: "Timeout for answer"
...